api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "API for LeFil"
authors = [
    {name = "Kamigaku",email = "kamigaku@gmail.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "structure @ file:///Code/LeFil/structure"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
